requirements.txt
line-bot-sdk
flask
beautifulsoup4
requests
pyquery
